+++ genai-web/packages/web/src/features/landing/components/ModelSelector.tsx
@@ -1,10 +1,11 @@
 import { CustomSelect } from '@/components/ui/CustomSelect';
 import { useSelectedModel } from '@/hooks/useSelectedModel';
-import { findModelDisplayNameByModelId, MODELS } from '@/models';
+import { useAvailableModels } from '@/hooks/useAvailableModels';
+import { findModelDisplayNameByModelId } from '@/models';
 
 export const ModelSelector = () => {
-  const { selectedModelId, setSelectedModelId } = useSelectedModel();
-  const { modelIds: availableModels } = MODELS;
+  const { selectedModelId, setSelectedModelId, availableModels } = useSelectedModel();
+  const { metadata } = useAvailableModels();
 
   return (
     <CustomSelect
@@ -13,7 +14,7 @@ export const ModelSelector = () => {
       value={selectedModelId}
       onChange={setSelectedModelId}
       options={availableModels.map((m) => {
-        return { value: m, label: findModelDisplayNameByModelId(m) };
+        return { value: m, label: findModelDisplayNameByModelId(m, metadata) };
       })}
     />
   );